add test to compare transaction root hashes to GETH results
